Cleanup
Banging my head against the wall with this one.

I cleaned up the project quite a bit, and unfortunately found a ton of bugs - some which originated from the original code.

It took quite a bit of refactoring but some major changes for now are:
-New node Euclidean distance calculations
-CheckNeighbour Function Revamped
-ReverseComparator changed to work properly now
-Better map usage, including edit mode
-Map fixes
-Navigation Fixes

The CPU version is pretty stable now.